requirements.txt
pytest==7.4.3
requests==2.31.0
pytest-html==4.1.1
python-dotenv==1.0.0
jsonschema==4.20.0
allure-pytest==2.13.2
pytest-xdist==3.3.1 